+++ features/step_definitions/auth_code_steps.rb
@@ -1,5 +1,5 @@
 O_AUTH_QUERY_PARAMS_WITH_CODE = {
-  redirect_uri:  "http://localhost:3000",
+  redirect_uri:  "http://example.org/",
   response_type: "code",
   scope:         "openid profile read",
   nonce:         "hello",
@@ -22,7 +22,7 @@ end
 When /^I parse the auth code and create a request to the token endpoint$/ do
   code = current_url[/(?<=code=)[^&]+/]
   post api_openid_connect_access_tokens_path, code: code,
-       redirect_uri: "http://localhost:3000", grant_type: "authorization_code",
+       redirect_uri: "http://example.org/", grant_type: "authorization_code",
        client_id: @client_id, client_secret: @client_secret
 end