+++ features/step_definitions/drawer_steps.rb
@@ -1,0 +1,5 @@
+And /^I click on "([^"]*)" in the drawer$/ do |txt|
+  within("#drawer") do
+    find_link(txt).trigger "click"
+  end
+end