+++ features/step_definitions/search_steps.rb
@@ -1,5 +1,5 @@
 When /^I enter "([^"]*)" in the search input$/ do |search_term|
-  fill_in "q", :with => search_term
+  find("input#q").native.send_keys(search_term)
 end
 
 When /^I click on the first search result$/ do
@@ -10,6 +10,13 @@ end
 
 When /^I press enter in the search input$/ do
   find("input#q").native.send_keys :return
+end
+
+When /^I search for "([^\"]*)"$/ do |search_term|
+  field = find_field("q")
+  fill_in "q", with: search_term
+  field.native.send_key(:enter)
+  expect(page).to have_content(search_term)
 end
 
 Then /^I should not see any search results$/ do