+++ features/support/matchers.rb
@@ -11,6 +11,18 @@ RSpec::Matchers.define :have_path do |expected|
   end
 end
 
+RSpec::Matchers.define :have_path_in do |expected|
+  match do |actual|
+    await_condition { expected.include? actual.current_path }
+  end
+
+  failure_message_for_should do |actual|
+    "expected #{actual.inspect} to have path in #{expected.inspect} but was #{actual.current_path.inspect}"
+  end
+  failure_message_for_should_not do |actual|
+    "expected #{actual.inspect} to not have path in #{expected.inspect} but it had"
+  end
+end
 
 RSpec::Matchers.define :have_value do |expected|
   match do |actual|
@@ -24,7 +36,6 @@ RSpec::Matchers.define :have_value do |expected|
     "expected #{actual.inspect} to not have value #{expected.inspect} but it had"
   end
 end
-
 
 def await_condition &condition
   start_time = Time.zone.now