+++ features/support/publishing_cuke_helpers.rb
@@ -5,7 +5,7 @@ module PublishingCukeHelpers
 
   def append_to_publisher(txt, input_selector='#status_message_fake_text')
     elem = find(input_selector)
-    elem.native.send_keys(' ' + txt)
+    elem.native.send_key(" #{txt}")
 
     # make sure the other text field got the new contents
     expect(find("#status_message_text", visible: false)).to have_value txt